+++ packages/slice-layout/src/templates/standard/index.js
@@ -11,8 +11,15 @@ const StandardSlice = ({ itemCount, renderItems }) => {
 
   return (
     <View style={styles.container}>
-      {renderItems(config({ itemCount })).map(item => (
-        <View key={item.props.id} style={styles.itemContainer}>
+      {renderItems(config({ itemCount })).map((item, index) => (
+        <View
+          key={item.props.id}
+          style={
+            index === itemCount - 1
+              ? styles.itemContainerWithoutBorders
+              : styles.itemContainer
+          }
+        >
           <View style={styles.item}>{item}</View>
         </View>
       ))}