+++ packages/utils/src/screen.base.js
@@ -1,26 +1,24 @@
 import { Dimensions } from "react-native";
-import {
-  tabletRowPadding,
-  tabletWidth,
-  tabletWidthMax
-} from "@times-components/styleguide";
+import { tabletRowPadding, tabletWidth } from "@times-components/styleguide";
 
 export const acceptedWidths = [
   320,
   440,
-  tabletWidth,
+  660,
   800,
   1080,
-  tabletWidthMax,
+  1280,
   1440,
   1670,
   1920,
-  2308,
-  2736
+  2308
 ];
 
-export const normaliseWidth = width => {
-  const nWidth = acceptedWidths.find(w => width <= w);
+// We want to ensure a small number of caches that are more frequently "warm"
+// so we limit the number of resolutions we will request for assets
+// across devices to a common set
+export const normaliseWidthForAssetRequestCache = widthInPixels => {
+  const nWidth = acceptedWidths.find(w => widthInPixels <= w);
 
   return nWidth || acceptedWidths[acceptedWidths.length - 1];
 };