+++ packages/utils/src/screen.js
@@ -1,10 +1,15 @@
 import { PixelRatio } from "react-native";
 
-import { acceptedWidths, normaliseWidth, screenWidth } from "./screen.base.js";
+import {
+  acceptedWidths,
+  normaliseWidthForAssetRequestCache,
+  screenWidth
+} from "./screen.base.js";
 
-export { acceptedWidths, normaliseWidth, screenWidth };
+export { acceptedWidths, normaliseWidthForAssetRequestCache, screenWidth };
 
 export const convertToPixels = points =>
   PixelRatio.getPixelSizeForLayoutSize(points);
 
-export const screenWidthInPixels = () => convertToPixels(screenWidth());
+export const screenWidthInPixels = isTablet =>
+  convertToPixels(screenWidth(isTablet));