+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.23](https://github.com/newsuk/times-components/compare/@times-components/ad@2.4.22...@times-components/ad@2.4.23) (2019-02-15)
+
+**Note:** Version bump only for package @times-components/ad
+
+
+
+
+
 ## [2.4.22](https://github.com/newsuk/times-components/compare/@times-components/ad@2.4.21...@times-components/ad@2.4.22) (2019-02-15)
 
 **Note:** Version bump only for package @times-components/ad