+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.21](https://github.com/newsuk/times-components/compare/@times-components/article-list@6.4.20...@times-components/article-list@6.4.21) (2019-02-15)
+
+
+### Bug Fixes
+
+* articles should not be larger than screen width ([#1708](https://github.com/newsuk/times-components/issues/1708)) ([48d851b](https://github.com/newsuk/times-components/commit/48d851b))
+
+
+
+
+
 ## [6.4.20](https://github.com/newsuk/times-components/compare/@times-components/article-list@6.4.19...@times-components/article-list@6.4.20) (2019-02-15)
 
 **Note:** Version bump only for package @times-components/article-list