+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.2](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.14.1...@times-components/article-magazine-standard@0.14.2) (2019-02-15)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [0.14.1](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.14.0...@times-components/article-magazine-standard@0.14.1) (2019-02-15)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard