+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@0.12.1...@times-components/article-skeleton@0.12.2) (2019-02-15)
+
+
+### Bug Fixes
+
+* articles should not be larger than screen width ([#1708](https://github.com/newsuk/times-components/issues/1708)) ([48d851b](https://github.com/newsuk/times-components/commit/48d851b))
+
+
+
+
+
 ## [0.12.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@0.12.0...@times-components/article-skeleton@0.12.1) (2019-02-15)
 
 **Note:** Version bump only for package @times-components/article-skeleton