+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.21](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.2.20...@times-components/video-label@2.2.21) (2019-02-15)
+
+**Note:** Version bump only for package @times-components/video-label
+
+
+
+
+
 ## [2.2.20](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.2.19...@times-components/video-label@2.2.20) (2019-02-15)
 
 **Note:** Version bump only for package @times-components/video-label